requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.8
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pillow==11.1.0
priority==2.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pypika-tortoise==0.3.2
pytz==2024.2
Quart==0.20.0
quart-cors==0.8.0
reportlab==4.2.5
sniffio==1.3.1
starlette==0.45.3
tortoise-orm==0.23.0
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
